refactor(visual-checks): extract VisualCheckStickyHeader FC
